# apps/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Home Sensor API",
    "description": "A collection for testing the Smart Home Sensor Management API",
    "version": "1.0.0"
  },
  "servers": [],
  "paths": {
...
    "/api/v1/sensors": {
      "get": {
        "tags": [
          "Get All Sensors"
        ],
        "summary": "Get All Sensors",
        "description": "Retrieve all sensors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "post": {
        "tags": [
          "Create Sensor"
        ],
        "summary": "Create Sensor",
        "description": "Create a new sensor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Living Room Temperature",
                  "type": "temperature",
                  "location": "Living Room",
                  "unit": "°C"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/sensors": {
      "get": {
        "tags": [
          "Get All Sensors v2"
        ],
        "summary": "Get All Sensors",
        "description": "Retrieve all sensors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "post": {
        "tags": [
          "Create Sensor"
        ],
        "summary": "Create Sensor",
        "description": "Create a new sensor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  }
}